refactor: handle messages in server struct
