+++ server/internal/listener/listener.go
@@ -1,134 +1,8 @@
-package internal
+package listener
 
-import (
-	"bufio"
-	"fmt"
-	"io"
-	"log"
-	"net"
-
-	"github.com/willcruse/kvdb/server/v2/internal"
-	storagebackend "github.com/willcruse/kvdb/server/v2/internal/storage-backend"
-	writelogger "github.com/willcruse/kvdb/server/v2/internal/write-logger"
-)
+import "net"
 
 // Listener is responsible for responding to incoming messages
 type Listener interface {
-	Listen() error
-}
-
-type TCPListener struct {
-	Address  string
-	Storage  storagebackend.StorageBackend
-	OpLogger writelogger.WriteOperationLogger
-}
-
-func (t *TCPListener) Listen() error {
-	ln, err := net.Listen("tcp", t.Address)
-	if err != nil {
-		return err
-	}
-
-	for {
-		conn, err := ln.Accept()
-		if err != nil {
-			log.Printf("tcp_listener: Error accepting connection. %v\n", err)
-			continue
-		}
-
-		go t.handleNetConn(conn)
-
-	}
-}
-
-func (t *TCPListener) handleNetConn(conn net.Conn) {
-	defer conn.Close()
-	reader := bufio.NewReader(conn)
-
-	for {
-		commandValue, err := reader.ReadByte()
-		if err != nil {
-			log.Printf("handler_net_conn: Failed to read command value from stream. %v\n", err)
-			return
-		}
-
-		commandValueInt := int(commandValue)
-		key, err := readString(reader)
-		if err != nil {
-			log.Printf("handler_net_conn: Error reading key from stream %v\n", err)
-			return
-		}
-
-		fmt.Printf("Command Value: %d\n", commandValueInt)
-
-		switch commandValueInt {
-		case commands.GET_COMMAND:
-			// handler get command
-			fmt.Printf("Fetching %s\n", key)
-			res, err := t.Storage.Get(key)
-			if err != nil {
-				return
-			}
-			fmt.Printf("Fetched %s -> %s\n", key, res)
-
-		case commands.SET_COMMAND:
-			value, err := readString(reader)
-			if err != nil {
-				log.Printf("handler_net_conn: Error reading value from stream %v\n", err)
-				return
-			}
-			fmt.Printf("Setting %s to %s\n", key, value)
-			err = t.Storage.Set(key, value)
-			if err != nil {
-				log.Printf("handler_net_conn: Error setting value %v\n", err)
-				return
-			}
-
-			err = t.OpLogger.LogSet(key, value)
-			if err != nil {
-				log.Printf("handler_net_conn: Error logging operation %v\n", err)
-				return
-			}
-			fmt.Printf("Set %s -> %s\n", key, value)
-
-		case commands.DELETE_COMMAND:
-			fmt.Printf("Deleting %s\n", key)
-			err = t.Storage.Delete(key)
-			if err != nil {
-				log.Printf("handler_net_conn: Error deleting value %v\n", err)
-				return
-			}
-			err = t.OpLogger.LogDelete(key)
-			if err != nil {
-				log.Printf("handler_net_conn: Error logging operation %v\n", err)
-				return
-			}
-			fmt.Printf("Delete %s\n", key)
-
-		default:
-			// TODO: handle unknown command
-		}
-	}
-
-}
-
-func readString(reader *bufio.Reader) (string, error) {
-	keyLength, err := reader.ReadByte()
-	if err != nil {
-		err = fmt.Errorf("handler_net_conn: Failed to read length from stream. %w", err)
-		return "", err
-	}
-
-	keyBuf := make([]byte, int(keyLength))
-	nRead, err := io.ReadFull(reader, keyBuf)
-	if err != nil {
-		err = fmt.Errorf("handler_net_conn: Failed to read value from stream. %w", err)
-		return "", err
-	}
-	if nRead != int(keyLength) {
-		err = fmt.Errorf("handler_net_conn: Incorrect number of bytes read from stream")
-		return "", err
-	}
-
-	return string(keyBuf), nil
+	Listen(connChan chan net.Conn) error
 }
